introducing dependency types
The dependency resolver can now differentiate between dependency types
The actual deptype is a bit field that sets flags like 'is mandatory',
'is an external dep', 'try other dep types, too'.

Functionality for this has been added to most modules/classes, particularly
to EbuildJobChannel and ebuild/depres, but not to the rule reader,
which means that there's currently no way to _use_ the dep types!

Also added an ErrorQueue implementation that saves some "if <has err_queue>"
checks. It's also able to unblock waiting queues, which helps to end
threads when an exception has been raised.