overlay/pkgdir/distroot replaces ./symlink
The functionality of Distroot/Distdir has been extended by support
for hard links and (potentially) file copies, thus the "symlink"
module has been (partially) rewritten and renamed to "distroot".

This commit adds transparent (i.e., manifest creation etc. doesn't have to
care about the configuration) access to DISTDIRs via
get_distdir(<ebuild_name>).
Adding files to a distdir is done with <distdir>.add (<package_file>).
Whether this results in a sym-, hardlink or a file copy in a temporary or
persistent directory is up to the "distdir strategy" and only known to
the "Distroot".