Package Rules: 'trace' action
The 'trace' action can be used to track changes done by package rules.

It has to implementations (action classes), depending on whether an arg is
supplied in the rule file or not:

* TraceAction (with arg), which adds the arg (typically a string) to a
  package's modified_by_package_rules variable when apply_action() is called

* MarkAsModifiedAction (without arg), which simply sets the
  modified_by_package_rules variable to True if it isn't already set.

Note that this is a testing/debugging feature and that there's currently no
(simple) method to access the trace variable.