distmap - track changes in DISTDIR
distmap is a persistent object (written to/read from file) that contains
information about files in distdir, e.g. name of the repo a file belongs to.
It also contains a checksum (sha256 as it's also used in Manifest files) that
can be used to check whether an ebuild has to be regenerated (without scanning
the entire overlay).

The file can (optionally) be bzip2- or gzip-compressed.