roverlay/depres: return depresult objects
Dependency resolution now creates depresult objects if a rule matches a
dependency string. Previously, it returned a 2-tuple (score,resolving_dep).

This is "necessary" for selfdep reduction (also see pseudocode in
overlay/root.py).

Additionally, this commit allows to define selfdeps that don't belong to the
default category ("sci-R") in rule files, either by OR-ing a deptype with
"#deptype <deptype>,selfdep" or by prefixing a single rule with "@selfdep".
The rule stub syntax ("zoo") is mostly unaffected by this change, but roverlay
is able to detect (non-)default-category selfdeps now.