roverlay/rpackage, config: add support for LICENSE
The description reader is now able to read the license field so that ebuild
creation can use it to create a $LICENSE variable (without processing it
further).

To achieve this, $PORTDIR/licenses/ is scanned for valid licenses. Additionally,
a license map file can be used to map unknown license strings to valid ones
(valid for portage in this context).