pkgdir: newmanifest
roverlay already creates checksums for package files since they're required for
the distmap. Using ebuildmanifest does not allow to re-use these checksums for
faster manifest file creation. In case of portagemanifest, it would (probably) be
possible.

This commit adds a mostly internal manifest file creation.
Compared with ebuildmanifest, it is A LOT faster and threadsafe, and in contrast
to portagemanifest, it does not rely on portage libs (which is hard to
maintain/verify (for me)) and is slightly faster.
Actually, newmanifest is meant to replace portagemanifest and become the default
implementation.

newmanifest falls back to "ebuild <file> manifest" for imported ebuilds since
SRC_URI parsing is not implemented.