roverlay/depres: non-greedy depres channel
The non-greedy channel processes all deps, whether successful or not. The
default behavior (EbuildJobChannel) is to stop if any mandatory dep cannot be
resolved. This is desirable when running overlay creation (it's faster),
but can hide unresolvable dep strings, which is not useful when writing dep
rules. Single-threaded depres shouldn't be affected, though, since the ebuild
channel has to wait until the resolver is done.

So, the typical use case for non-greedy depres channels is when dealing with
partial results (which overlay creation doesn't).