R package description reader: fix race condition
This commit fixes a race condition concerning WRITE_DESCFILES_DIR. Setting it at
module import time doesn't work, because roverlay's config is not initialized,
so WRITE_DESCFILES_DIR was always None (=disabled).

DescriptionReader now initializes this variable when the first object is created
(calls setup actions in __new__()). This, however, means that an object has to
be created before the variable is accessible (!).
Additionally, some common (not object-specific) variables are now available
as class-wide variables (no longer bound to self or locals()).