depres, dep env: provide a ref to the package info
DepEnv objects can now (optionally) contain a reference to "their" PackageInfo
object. This allows the dynamic selfdep pool to decide which rules to try (and
in which order) based on the package being processed.
(I've chosen a (weak) reference because DepEnv objects should not keep
PackageInfos in memory.)

The actual PackageInfo can then be accessed via <DepEnv>.get_package_info().