write stats to a text file
roverlay used to write its stats to a database managed by using rrdtool. This
turns out not to be useful because
(a) extra dependency on rrdtool
(b) rrdtool's features are not used - only the last stats are queried
(c) "heavy"/lengthy implementation

roverlay now writes its stats to a plain text file, which is sufficient (for
now).

The rrdtool implementation is going to be (re-)moved.