roverlay/config/, ConfigTree: get options by name
Added get_by_name() to the ConfigTree, which is similar to get(), but takes an
option name (used in config files, e.g. OVERLAY_DIR) instead of a config path
(used in the code, e.g. ['OVERLAY','dir']) as key.
query_by_name() can be used to get multiple options and return them as dict.

Might be useful for scripting.