roverlay: add webgui/ (WIP)
A django web interface for roverlay.

The main motivation is manageable dependency rule creation/editing
(for example, feed the webgui with a list of unresolvable dep strings,
create rules for some of them (leaving the rest as TODO),
let someone else step in and correct/create rules, ...).

See webgui/README for its current state (in short, the actual web interface is
TODO, whereas the (depres) database part is mostly done).