deprules: remove virtual/jdk:1.5 rule (workaround)
SLOT-relative dependency rules do not take the version operator (<, >, ==, !=,
<=, >=) into account, which is a logic/design error because dep strings like
"Java (>= 5.0)" can be interpreted as
(a) any version within the 5.0 release series
(b) any version greater or equal than 5.0, e.g. 7.0

SLOT rules always assume (a), but (b) is equally possible - depending on
the R package in question as well as the respective rule.

This commit is a *workaround* for a single dep rule.
It removes the SLOT rule for virtual/jdk:1.5, which is not available in
the main tree and some packages affected by this rule (RWeka, RWekajars) have
been confirmed to work with more recent java releases.
Effectively, this means that all "Java <somehow> 5.0" dep strings get resolved
as ">=virtual/jdk-1.5" now, which breaks packages that really depend on java
1.5 (should be solvable by installing it manually).

All other SLOT rules (=3) are affected by this issue, too.

Needs to be fixed code-wise.

Reported-By: Randall Wald <rdwald@gmail.com>