util/dictwalk: replace dictmerge()
dictmerge() was undocumented and had counter-intuitive defaults
for get-value/get-key (get-value returned a "scalar", get-key a list).
Additionally, the function's name was quite ambiguous (e.g. one
could expect that this function merges N dicts, which is not the case).

This commit adds accumulate_key_value_list() as more properly-defined
replacement.
Also adds a helper function for merging a list of pairs (key,value 2-tuples),
which is the common use case (actually, there's only one "consumer" of this
function).